assistbuddy/requirements.txt
# Core deep learning
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0
bitsandbytes>=0.41.0
peft>=0.6.0
datasets>=2.14.0

# Vision & OCR
opencv-python>=4.8.0
Pillow>=10.0.0
pytesseract>=0.3.10
easyocr>=1.7.0
timm>=0.9.0

# Document processing
pdfplumber>=0.10.0
PyPDF2>=3.0.0
python-docx>=1.0.0
openpyxl>=3.1.0
pandas>=2.0.0

# Audio & Video
whisper>=1.1.10
openai-whisper>=20231117
ffmpeg-python>=0.2.0
librosa>=0.10.0

# Web scraping
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0

# PII detection & NLP
spacy>=3.7.0
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
regex>=2023.10.0

# Training utilities
tensorboard>=2.14.0
wandb>=0.15.0
tqdm>=4.66.0
numpy>=1.24.0
scipy>=1.11.0

# Inference & Demo
gradio>=4.0.0
flask>=3.0.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Evaluation
rouge-score>=0.1.2
sacrebleu>=2.3.0
nltk>=3.8.0

# Google Colab specific
google-colab
